MODE-1745 MODE-1739 Changed how Delta impls are created
Recent changes changed the way Delta implementations are created.
In particular, the SchematicEntryWholeDelta instances were created
with a clone of the document. Since subsequent changes were made
to the original document, the SchematicEntryWholeDelta didn't include
these subsequent changes and they were lost, causing various exceptions.

Also changed some of the trace logging messages on both Delta
implementations to be more useful.

I'm able to not only successfully run a full integration build,
but I'm able to successfully run the test cases for both MODE-1745
and MODE-1739 using Infinispan 5.1.2.FINAL or 5.1.8.Final.
Occassionally when running the MODE-1739 tests I do get one or two
NodeNotFoundInParentException (see MODE-1739 for details) early
on in the test (unable to find a path of an existing node when
checking permissions), but it simply fails that request and continues.
This error often happens almost immedidately in the test and is
rarely repeated more than once.

Note that in addition to this problem, the tests are not able to use
Infinispan 5.2 -- that's because we're using the SchematicEntryDelta
(a diff-based impl), and we haven't yet fixed the problem there.