MODE-1762 Added code to ensure the sequencers always close the Binary's stream
Our Binary values always return a stream that closes itself when fully-read.
However, if any of the sequencers do not read the entire stream (e.g.,
intentionally or if they experience an exception while processing),
the stream might not be closed.

These changes ensure that the stream is always closed. All sequencer
implementations where checked, and only those that were not always
closing the stream in a finally block were modified/corrected.

All tests pass with these changes.