MODE-1769 Added support for async cache store
A bug in Infinispan caused a fatal exception when attempting to use
an asynchronous cache store because the IsMarshallableInterceptor
was incorrectly determining that SchematicEntryLiteral could not
be marshalled when in fact it does have a valid Infinispan externalizer.

The workaround is to remove the IsMarshallableInterceptor from the
cache's interceptor stack ONLY when the async cache store is used
and only on versions of Infinispan *beore* 5.2.0.Final (including
5.2.0.CR1). See ISPN-2748 for details, which should be fixed in
5.2.0.Final.

Added a test case that used an asynchronous cache store to first
replicate the problem and then verify it was fixed. (Also used
a separate project to verify it works with multiple versions of
Infinispan, including 5.1.2.FINAL, 5.1.8.Final, and 5.2.0.CR1).