MODE-1770 Added PESSIMISTIC locking to the default, in-memory Infinispan configuration
The LocalEnvironment class and its createDefaultConfiguration() method
are used to create the default cache configurations for in-memory content.
This method does set up the transactional behavior, but does not set
the locking. Therefore, I think it is prudent to add code to this
method that sets up the LockingMode of PESSIMISTIC.

Verified that this has no negative effect on the tests.