MODE-1790 Corrected InfinispanBinaryStore to remove temporary files
The InfinispanBinaryStore was creating temporary files to determine
the SHA-1 of the file, which is used for the cache key and thus
must be known before doing anything with the cache. But it was not
deleting these temporary files when finished.

This change removes the file in the existing finally block of the
'storeValue' method. It also uses local variables for the size
and lastModified timestamp of the file.