MODE-1795 Importing XML should map default XML namespace
When importing XML, any XML namespaces should be used, even if the
prefix for the XML namespace matches an existing namespace prefix.
Per Section 11.1 of the JSR-283 specification, an existing JCR namespace
should be used if its URL matches the XML namespace URI. If
there is no existing JCR namespace matching the XML namespace URI,
then the namespace should be registered using the XML namespace prefix
if not already used or a generated prefix if the XML namespace prefix
is already used.

A new unit test case was added to verify the behavior.