MODE-1852 Correct deserialization of events over JGroups channel
JGroups needs to know about the classloader that loads the ModeShape
event classes. The easiest way to do this is to extend ObjectInputStream
to supply and use a classloader for deserialization. This is especially
important in AS7/EAP and OSGi, where a single classloader is not used
for all the components.