MODE-1865 Additional changes and fixes
The test case added in previous commits (for this issue) was not cleaning
up the 300 nodes it was creating and using for the queries, thus causing
failures in many subsequent tests. This was corrected, and the test case
simplified. The changes to QueryProcessor were also incomplete and did
not always return just one row when using "LIMIT 1", so this was corrected.