MODE-1878 Corrected determination of correct SNS child node definitions
When a node with a child is persisted, and then a new child node with the
same name as the existing child is added to the parent, and when the
parent node type contains more than one applicable residual child node definition,
ModeShape incorrected determined that no child node definition could be
found for the new SNS child.

Two new test cases were added to replicate this scenario, and the
validation logic was corrected.