MODE-1880 Minor improvement to generation of IDs inside the file system connector
When the file system connector writes nodes to the file system, a warning
is logged for every node (other than 'jcr:content' nodes) contain
a namespace in their name. This is because the file system connector
turns only the local part of the node name into the file/folder name.