MODE-1873 Corrected how subquery results are converted to criteria
When a subquery selected a pseudo-column (like 'jcr:path'), the query
result contained Path instances. These values could not be successfully
used in criteria, since they were not converted to the correct value type
expected in the dependent query. Since only pseudo-columns were a problem
(and really only the 'jcr:path' and 'jcr:name' pseud-columns since they
return Path and Name values, respectively), the code that grabs the
subquery results was modified to convert these non-standard types
to strings.

A test case was added to the text sequencer test case, since that was
the best place for replicating the problem query (which used the
'mode:derivedFrom' property created during the sequencing operation).
However, once the problem was identified, I created a simple test
case within the JcrQueryManagerTests.