MODE-1876 Removed expensive ops to allow reindexing large repositories
Recently the reindexing operations were changed so that upon "IF_MISSING"
any already-indexed node would be skipped. However, this was not really
the intention of the "IF_MISSING" option, and the logic to determine which
nodes were already indexed was quite expensive. Therefore, the logic
was changed to reindex the entire repository's content whenever the
indexes were missing (e.g., either empty or not yet created).