MODE-1896 Deprecated the "transactionManagerLookup" field
This field in repository configuration files has not been used for some
time, but at the time it was not properly documented in the schema as
being "deprecated". This has been done, and any use of this field in
our test configurations have been removed.

Note that the RepositoryConfiguration constructors now log warning messages
when this or any other deprecated fields are used.