MODE-1937 Corrected the Java sequencer's generated method nodes
Nodes representing methods were named with the method name and parameter
names, but this was changed to be closer to the standard Java signature
of method name and parameter types. This makes it possible to properly
support overloaded methods and constructors.

Several of the existing tests were modified to use overloaded
constructors and methods, and all pass with these changes. Note that
no changes in the node types were required, though the sequencer's
generated content is now slightly different (only in terms of the method
and constructor node names).