MODE-1946 Corrected WebDAV behavior when putting file with OS X
The OS X Finder application is used on OS X to mount a WebDAV service (e.g.,
a repository) as a networked drive. However, the behavior of Finder
is verbose in that it sends multiple PUT operations when copying a
single file into the WebDAV area. The first PUT has no content,
and is followed by a LOCK, a second PUT (with the content), and
and UNLOCK. ModeShape's WebDAV service was not expecting this pattern
and was not writing the second content.

The fix was relatively gross, since it entails OS-specific checks.
But it does work, and all tests do pass.