MODE-1444: Store binary files using Cassandra database
Two mechanisms were added to add support for storing binary values in
Cassandra. The first was via the JDBC binary store, with the Cassandra
JDBC driver as an option. The second was via a new CassandraCacheStore
class that makes it a little easier and more straightforward than the
JDBC binary store approach.

Note that in order to use either approach, the embedding applications
must include the Cassandra libraries on the classpath; ModeShape simply
expects them to be provided.